HID: nintendo: use slots for led patterns
Previously, the leds pattern would just increment with every controller
connected. This wouldn't take into consideration when controllers are
disconnected. The same controller could be connected and disconnected
with the pattern increasing player count each time.

This changes it so now a bitmask is used defining the free controller
slots for each "player". When a controller is connected, it searches
for a bit in the bitmask and acquires that "slot" for it's led pattern.

When a controller is disconnected, it sets the position back to 'free'
allowing for a new controller that is to be connected to obtain that
slot.

Signed-off-by: Ryan McClelland <rymcclel@gmail.com>